+++ src/main/java/ua/roman/flats/repositories/name/NameVoivodeshipRepository.java
@@ -3,5 +3,5 @@ package ua.roman.flats.repositories.name;
 import org.springframework.data.repository.CrudRepository;
 import ua.roman.flats.domain.name.NameVoivodeship;
 
-public interface NameVoivodeshipRepository extends CrudRepository<NameVoivodeship, Long> {
+public interface NameVoivodeshipRepository extends CrudRepository<NameVoivodeship, Integer> {
 }
